Conexion a la Base de Datos funcionando correctamente

+++ WebApplication1/WebApplication1/Models/agenda.cs
@@ -16,7 +16,7 @@ namespace WebApplication1.Models
     {
         public int id { get; set; }
         public string nombre { get; set; }
-        public Nullable<decimal> celular { get; set; }
+        public Nullable<long> celular { get; set; }
         public string email { get; set; }
         public string direccion { get; set; }
     }
